back/requirements.txt
Flask>=3.0.0
numpy>=1.20.0
Flask-MySQL>=1.5.2
pymysql==1.1.0
Flask-CORS>=3.0.10
python-dotenv==0.19.0
Flask-JWT-Extended>=4.3.1
Flask-Bcrypt>=1.0.1
# Werkzeug==0.16.1
# MarkupSafe==1.1.1
# Jinja2==2.11.3
PyJWT>=2.8.0
